Use memcpy instead of a char-by-char loop to fix the mysterious verification bug

+++ include/device/AHCIDevice.h
@@ -9,9 +9,9 @@ namespace Thorn::Device {
 		AHCIDevice() = delete;
 		AHCIDevice(AHCI::Port *port_): port(port_) {}
 
-		virtual int read(void *buffer, size_t size, off_t offset) override;
-		virtual int write(const void *buffer, size_t size, off_t offset) override;
-		virtual int clear(off_t offset, size_t size) override;
+		virtual int read(void *buffer, size_t size, size_t offset) override;
+		virtual int write(const void *buffer, size_t size, size_t offset) override;
+		virtual int clear(size_t offset, size_t size) override;
 		virtual std::string getName() const override;
 	};
 }
