+++ src/device/IDEDevice.cpp
@@ -3,19 +3,19 @@
 #include "hardware/IDE.h"
 
 namespace Thorn::Device {
-	int IDEDevice::read(void *buffer, size_t size, off_t offset) {
+	int IDEDevice::read(void *buffer, size_t size, size_t offset) {
 		if (offset % IDE::SECTOR_SIZE == 0 && size % IDE::SECTOR_SIZE == 0)
 			return IDE::readSectors(ideID, size / IDE::SECTOR_SIZE, offset / IDE::SECTOR_SIZE, (char *) buffer);
 		return IDE::readBytes(ideID, size, offset, buffer);
 	}
 
-	int IDEDevice::write(const void *buffer, size_t size, off_t offset) {
+	int IDEDevice::write(const void *buffer, size_t size, size_t offset) {
 		if (offset % IDE::SECTOR_SIZE == 0 && size % IDE::SECTOR_SIZE == 0)
 			return IDE::writeSectors(ideID, size / IDE::SECTOR_SIZE, offset / IDE::SECTOR_SIZE, buffer);
 		return IDE::writeBytes(ideID, size, offset, buffer);
 	}
 
-	int IDEDevice::clear(off_t offset, size_t size) {
+	int IDEDevice::clear(size_t offset, size_t size) {
 		if (offset % IDE::SECTOR_SIZE == 0 && size % IDE::SECTOR_SIZE == 0) {
 			char buffer[IDE::SECTOR_SIZE] = {0};
 			int status;
